Clean up HString methods
Move stuff that makes HString behave like a basic array to a base class.
