+++ src/BaseHString.php
@@ -1,0 +1,205 @@
+<?php
+namespace Haystack;
+
+use Haystack\Helpers\Helper;
+
+abstract class BaseHString implements \Iterator, \ArrayAccess, \Serializable, \Countable
+{
+    protected $string;
+    private $ptr; // pointer for iterating through $string
+
+    public function __construct($string = "")
+    {
+        if (is_scalar($string) || $string instanceof HString) {
+            $this->string = (string) $string;
+            $this->rewind();
+        } elseif (is_null($string)) {
+            $this->string = null;
+        } else {
+            throw new \ErrorException(sprintf("%s is not a proper String", Helper::getType($string)));
+        }
+    }
+
+    /**
+     * @return string
+     */
+    public function __toString()
+    {
+        return sprintf($this->string);
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.0.0)<br/>
+     * Whether a offset exists
+     *
+     * @link http://php.net/manual/en/arrayaccess.offsetexists.php
+     * @param mixed $offset <p>
+     *                      An offset to check for.
+     *                      </p>
+     * @return boolean true on success or false on failure.
+     *                      </p>
+     *                      <p>
+     *                      The return value will be casted to boolean if non-boolean was returned.
+     */
+    public function offsetExists($offset)
+    {
+        return isset($this->string[$offset]);
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.0.0)<br/>
+     * Offset to retrieve
+     *
+     * @link http://php.net/manual/en/arrayaccess.offsetget.php
+     * @param mixed $offset <p>
+     *                      The offset to retrieve.
+     *                      </p>
+     * @return mixed Can return all value types.
+     */
+    public function offsetGet($offset)
+    {
+        return $this->string[$offset];
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.0.0)<br/>
+     * Offset to set
+     *
+     * @link http://php.net/manual/en/arrayaccess.offsetset.php
+     * @param mixed $offset <p>
+     *                      The offset to assign the value to.
+     *                      </p>
+     * @param mixed $value  <p>
+     *                      The value to set.
+     *                      </p>
+     * @return void
+     */
+    public function offsetSet($offset, $value)
+    {
+        $this->string[$offset] = $value;
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.0.0)<br/>
+     * Offset to unset
+     *
+     * @link http://php.net/manual/en/arrayaccess.offsetunset.php
+     * @param mixed $offset <p>
+     *                      The offset to unset.
+     *                      </p>
+     * @return void
+     */
+    public function offsetUnset($offset)
+    {
+        $this->string[$offset] = null;
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.1.0)<br/>
+     * String representation of object
+     *
+     * @link http://php.net/manual/en/serializable.serialize.php
+     * @return string the string representation of the object or null
+     */
+    public function serialize()
+    {
+        return serialize($this->toString());
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.1.0)<br/>
+     * Constructs the object
+     *
+     * @link http://php.net/manual/en/serializable.unserialize.php
+     * @param string $value <p>
+     *                           The string representation of the object.
+     *                           </p>
+     * @return void
+     */
+    public function unserialize($value)
+    {
+        if (is_scalar($value)) {
+            $this->string = unserialize($value);
+        } elseif (is_null($value)) {
+            $this->string = null;
+        } else {
+            throw new \InvalidArgumentException(sprintf("HString cannot unserialize a %s", Helper::getType($value)));
+        }
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.1.0)<br/>
+     * Count elements of an object
+     *
+     * @link http://php.net/manual/en/countable.count.php
+     * @return int The custom count as an integer.
+     *       </p>
+     *       <p>
+     *       The return value is cast to an integer.
+     */
+    public function count()
+    {
+        return strlen($this->string);
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.0.0)<br/>
+     * Return the current element
+     *
+     * @link http://php.net/manual/en/iterator.current.php
+     * @return mixed Can return any type.
+     */
+    public function current()
+    {
+        return $this->string[$this->ptr];
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.0.0)<br/>
+     * Move forward to next element
+     *
+     * @link http://php.net/manual/en/iterator.next.php
+     * @return void Any returned value is ignored.
+     */
+    public function next()
+    {
+        ++$this->ptr;
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.0.0)<br/>
+     * Return the key of the current element
+     *
+     * @link http://php.net/manual/en/iterator.key.php
+     * @return integer scalar on success, or null on failure.
+     */
+    public function key()
+    {
+        return $this->ptr;
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.0.0)<br/>
+     * Checks if current position is valid
+     *
+     * @link http://php.net/manual/en/iterator.valid.php
+     * @return boolean The return value will be casted to boolean and then evaluated.
+     *       Returns true on success or false on failure.
+     */
+    public function valid()
+    {
+        return $this->ptr < $this->count();
+    }
+
+    /**
+     * (PHP 5 &gt;= 5.0.0)<br/>
+     * Rewind the Iterator to the first element
+     *
+     * @link http://php.net/manual/en/iterator.rewind.php
+     * @return void Any returned value is ignored.
+     */
+    public function rewind()
+    {
+        $this->ptr = 0;
+    }
+}
